Gemini 음성 연결, 길찾기 중 첫번 째 만 네비게이션
