+++ requirements.txt
@@ -1,3 +1,4 @@
 google-generativeai
 geocoder
 shapely
+python-dotenv
